Move express setup into another module for testing.
